let cargo install elba command include +nightly
The `cargo` command to install elba currently does not have the `+nightly` flag. A careless reader of the installation guide, like myself, could miss the instruction to use a nightly version of Rust.

With this change the command specifies the dependency on nightly right where the action happens. I probably would still miss it ;-) but at least when copied, things can't go wrong.